LayoutTests/fast/canvas/webgl/texImage2D-video-flipY-false-expected.txt

Setting video src.
Video can play.
Checking the canvas pixels.
PASS: Bottom edge is red.
PASS: Top edge is blue.
